+++ UnicodeMazeRunner.cpp
@@ -6,7 +6,7 @@
 #include <cmath>
 #include <array>
 #include <ahadmin.h>
-#include "MapCreator.h"
+#include "./CommonLib/MapCreator.h"
 
 using namespace std;
